+++ types/styled-components/index.d.ts
@@ -12,7 +12,6 @@
 //                 Maciej Goszczycki <https://github.com/mgoszcz2>
 //                 Danilo Fuchs <https://github.com/danilofuchs>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
-// TypeScript Version: 2.9
 
 // forward declarations
 declare global {
@@ -96,11 +95,11 @@ type StyledComponentPropsWithAs<
 export type FalseyValue = undefined | null | false;
 export type Interpolation<P> =
     | InterpolationValue
-    | FlattenInterpolation<P>
-    | InterpolationFunction<P>;
-// must be an interface to be self-referential
-export interface FlattenInterpolation<P>
-    extends ReadonlyArray<Interpolation<P>> {}
+    | InterpolationFunction<P>
+    | FlattenInterpolation<P>;
+// cannot be made a self-referential interface, breaks WithPropNested
+// see https://github.com/microsoft/TypeScript/issues/34796
+export type FlattenInterpolation<P> = ReadonlyArray<Interpolation<P>>;
 export type InterpolationValue =
     | string
     | number
@@ -111,9 +110,7 @@ export type InterpolationValue =
 export type SimpleInterpolation =
     | InterpolationValue
     | FlattenSimpleInterpolation;
-// must be an interface to be self-referential
-export interface FlattenSimpleInterpolation
-    extends ReadonlyArray<SimpleInterpolation> {}
+export type FlattenSimpleInterpolation = ReadonlyArray<SimpleInterpolation>;
 
 export type InterpolationFunction<P> = (props: P) => Interpolation<P>;
 
@@ -148,7 +145,7 @@ type ForwardRefExoticBase<P> = Pick<
 // Config to be used with withConfig
 export interface StyledConfig<O extends object = {}> {
     // TODO: Add all types from the original StyledComponentWrapperProperties
-    shouldForwardProp?: (prop: keyof O, defaultValidatorFn: (prop: keyof O) => boolean) => boolean;
+    shouldForwardProp?: (prop: keyof O, defaultValidatorFn: ((prop: keyof O) => boolean)) => boolean;
 }
 
 // extracts React defaultProps
@@ -175,29 +172,12 @@ export interface StyledComponentBase<
     A extends keyof any = never
 > extends ForwardRefExoticBase<StyledComponentProps<C, T, O, A>> {
     // add our own fake call signature to implement the polymorphic 'as' prop
-    // NOTE: TS <3.2 will refuse to infer the generic and this component becomes impossible to use in JSX
-    // just the presence of the overload is enough to break JSX
-    //
-    // TODO (TypeScript 3.2): actually makes the 'as' prop polymorphic
-    // (
-    //     props: StyledComponentProps<C, T, O, A> & { as?: never }
-    //   ): React.ReactElement<StyledComponentProps<C, T, O, A>>
-    // <AsC extends keyof JSX.IntrinsicElements | React.ComponentType<any> = C>(
-    //   props: StyledComponentPropsWithAs<AsC, T, O, A>
-    // ): React.ReactElement<StyledComponentPropsWithAs<AsC, T, O, A>>
-
-    // TODO (TypeScript 3.2): delete this overload
     (
-        props: StyledComponentProps<C, T, O, A> & {
-            /**
-             * Typing Note: prefer using .withComponent for now as it is actually type-safe.
-             *
-             * String types need to be cast to themselves to become literal types (as={'a' as 'a'}).
-             */
-            as?: keyof JSX.IntrinsicElements | React.ComponentType<any>;
-            forwardedAs?: keyof JSX.IntrinsicElements | React.ComponentType<any>;
-        }
-    ): React.ReactElement<StyledComponentProps<C, T, O, A>>;
+        props: StyledComponentProps<C, T, O, A> & { as?: never, forwardedAs?: never }
+      ): React.ReactElement<StyledComponentProps<C, T, O, A>>;
+    <AsC extends keyof JSX.IntrinsicElements | React.ComponentType<any> = C>(
+      props: StyledComponentPropsWithAs<AsC, T, O, A>
+    ): React.ReactElement<StyledComponentPropsWithAs<AsC, T, O, A>>;
 
     withComponent<WithC extends AnyStyledComponent>(
         component: WithC
@@ -465,14 +445,12 @@ export class ServerStyleSheet {
 
 export type StylisPlugin = (
     context: number,
-    content: string,
     selector: string[],
     parent: string[],
+    content: string,
     line: number,
     column: number,
     length: number,
-    at: number,
-    depth: number
 ) => string | void;
 
 export interface StyleSheetManagerProps {
