+++ types/styled-components/package.json
@@ -5,9 +5,9 @@
     },
     "types": "index",
     "typesVersions": {
-        ">=3.7.0-0": {
+        "<=3.6": {
             "*": [
-                "ts3.7/*"
+                "ts3.6/*"
             ]
         }
     }
