+++ types/styled-components/ts3.6/test/index.tsx
@@ -18,8 +18,10 @@ import styled, {
     FlattenSimpleInterpolation,
     SimpleInterpolation,
     FlattenInterpolation
-} from "styled-components";
-import {} from "styled-components/cssprop";
+// tslint:disable-next-line:no-relative-import-in-test
+} from "..";
+// tslint:disable-next-line:no-relative-import-in-test
+import {} from "../cssprop";
 
 /**
  * general usage
@@ -340,7 +342,6 @@ const AttrsWithOnlyNewProps = styled.h2.attrs({ as: "h1" })`
 
 const AttrsInputExtra = styled(AttrsInput).attrs({ autoComplete: "off" })``;
 <AttrsInputExtra />;
-
 /**
  * withConfig
  */
@@ -475,10 +476,11 @@ class MyComponent extends React.Component<ThemeProps<{}>> {
 
 const ThemedMyComponent = withTheme(MyComponent);
 
-<ThemedMyComponent ref={ref => {
-    // $ExpectType MyComponent | null
-    ref;
-}}/>;
+// TODO: passes in TS@3.1, not in TS@3.0
+// <ThemedMyComponent ref={ref => {
+//     // $ExpectType MyComponent | null
+//     ref;
+// }}/>;
 const themedRef = React.createRef<MyComponent>();
 <ThemedMyComponent ref={themedRef} />;
 
@@ -680,7 +682,9 @@ class Test2Container extends React.Component<Test2ContainerProps> {
 }
 
 const containerTest = (
-    <StyledTestContainer as={Test2Container} type='foo' />
+    // TODO (TypeScript 3.2): once the polymorphic overload is un-commented-out this should be the correct test
+    // <StyledTestContainer as={Test2Container} type='foo' />
+    <StyledTestContainer as={Test2Container} size="small" />
 );
 
 // 4.0 refs
@@ -723,7 +727,8 @@ async function typedThemes() {
         createGlobalStyle,
         ThemeProvider,
         ThemeConsumer
-    } = (await import("styled-components")) as any as ThemedStyledComponentsModule<
+    // tslint:disable-next-line:no-relative-import-in-test
+    } = (await import("..")) as any as ThemedStyledComponentsModule<
         typeof theme
     >;
 
@@ -780,22 +785,6 @@ async function typedThemes() {
         ${themedCssWithNesting}
     `;
 
-    const WithProp = styled.div`
-        ${({ ok, theme: { color } }: { ok: boolean; theme: typeof theme }) =>
-            ok &&
-            css`
-                color: ${color};
-            `}
-    `;
-
-    const WithPropNested = styled.div`
-        ${({ ok }: { ok: boolean }) =>
-            ok &&
-            css`
-                color: ${({ theme: { color } }) => color};
-            `}
-    `;
-
     return (
         <ThemeProvider theme={theme}>
             <>
@@ -810,15 +799,14 @@ async function typedThemes() {
                         return theme.color;
                     }}
                 </ThemeConsumer>
-                <WithProp ok />
-                <WithPropNested ok />
             </>
         </ThemeProvider>
     );
 }
 
 async function reexportCompatibility() {
-    const sc = await import("styled-components");
+    // tslint:disable-next-line:no-relative-import-in-test
+    const sc = await import("..");
     const themed = sc as ThemedStyledComponentsModule<any>;
 
     let { ...scExports } = sc;
@@ -839,10 +827,12 @@ async function themeAugmentation() {
         accent: string;
     }
 
-    const base = (await import("styled-components")) as any as ThemedStyledComponentsModule<
+    // tslint:disable-next-line:no-relative-import-in-test
+    const base = (await import("..")) as any as ThemedStyledComponentsModule<
         BaseTheme
     >;
-    const extra = (await import("styled-components")) as any as ThemedStyledComponentsModule<
+    // tslint:disable-next-line:no-relative-import-in-test
+    const extra = (await import("..")) as any as ThemedStyledComponentsModule<
         ExtraTheme,
         BaseTheme
     >;
@@ -1055,7 +1045,8 @@ function validateDefaultProps() {
         color: red
     `;
 
-    <MyComponent requiredProp />;
+    // this test is failing in TS 2.9 but not in 3.0
+    // <MyComponent requiredProp />;
 
     <StyledComponent requiredProp optionalProp="x" />;
 
@@ -1076,7 +1067,8 @@ function validateDefaultProps() {
         { requiredProp: true }
     );
 
-    <OtherStyledComponent />;
+    // this test is failing in TS 3.1 but not in 3.2
+    // <OtherStyledComponent />;
 
     <OtherStyledComponent requiredProp="1" />; // $ExpectError
 }
@@ -1115,7 +1107,7 @@ function staticPropertyPassthrough() {
     interface BProps { b?: string; }
     interface BState { b?: string; }
     class A extends React.Component<AProps> {}
-    class B extends React.Component {
+    class B extends React.Component<BProps, BState> {
         static A = A;
         PUBLIC = 'PUBIC_VAL';
         static F = (props: BProps, state: BState) => props && state;
@@ -1123,21 +1115,14 @@ function staticPropertyPassthrough() {
             return state;
         }
     }
-    // Test FunctionComponent as well which can't be tested in <= TS 3.0
-    const C: React.FC & { A: typeof A; F: () => void } = () => <div></div>;
-    C.A = A;
-    C.F = () => {};
     const StyledB = styled(B)``;
-    const StyledC = styled(C)``;
     <StyledB.A />; // $ExpectError
     <StyledB.A a='a' />; // $ExpectError
     <StyledB.A a={0} />;
     StyledB.PUBLIC; // $ExpectError
     StyledB.componentDidMount(); // $ExpectError
-    StyledB.F({ b: 'b' } , {  b: 'b' });
+    StyledB.F({ b: 'b' } , { b: 'b' });
     StyledB.getDerivedStateFromProps({ b: 'b' } , { b: 'b' }); // $ExpectError
-    <StyledC.A a={0} />;
-    StyledC.F();
 }
 
 function unionTest() {
