+++ types/styled-components/ts3.6/tslint.json
